Remove non-working colored-output code
Ideally we'd print the method count in green, yellow, or red, depending
on the number, but our attempt at using undocumented Gradle interfaces
has never worked (it seems that setting a color via System.setProperty
is only effective during Gradle initialization, well before we get a
chance to set it).

Until we figure out a way to get this working, better to just kill the
code.