Remove `exportAsCSV`, print the summary every time
I'm unhappy with the proliferation of config options.  None of them (except `includeClassNames`, `printAsTree`, and `verbose`) are particularly useful, especially `exportAsCSV`.  The file it generates, however, is unquestionably useful.  We've had several user issues where summary counts were needed, and having a dedicated output file is far better than handing out awk snippets.

So - we kill the config knob and make the API simpler, and retain the useful part.  Along with that, the plugin task is cleaned up quite a bit as well.