Revert #32 and remove `@Input` from apkOrDex
This fixes a bug wherein the dexcount task fails if the variant's output
is renamed after the dexcount plugin is applied and prior to task execution.

This also opts dexcount out of Gradle's incremental compilation cache.  The
proximate cause of this bug is that we're holding a stale `java.io.File`
object; the cure is to instead hold a reference to the variant's output.
This doesn't work with incremental compliation, because variant output
implementations aren't `Serializable` (noted in #32), and Gradle depends
on task state being serializable.  This mismatch means that dexcount will
appear to randomly crash.

The least-bad solution is to just remove the task-input annotation, and
run the count task every time.  Not great, but better than not working.

Fixes #40 and #46.