Warn if Gradle is being run on Java 7 or below. (#120)
Android build tools are starting to be compiled with JDK 8, which means
that when invoking external tools (e.g. `dx` for counting .aar files),
we can't guarantee that we won't crash with a version-mismatch error.

Since the state of these tools is beyond our control, we may as well
strongly urge people to move to Java 8 for building.  At the very
least we need to disclaim responsibility if they don't, and that's what
this commit does.