Remove output annotations to opt out of Gradle caching
As it turns out, tasks with outputs but no inputs are considered by
Gradle to be up-to-date based on the outputs alone, meaning that you
can build, change your code, then rebuild, and not see updated counts.

This change removes the output annotations, fixing the staleness
problem - but also, opting out of Gradle's incremental builds entirely.

Fixes #131.