Replace modern 'InputDirectory' with 'InputFile'
Somewhere between Android Studio 3.0 alpha1 and beta6, the AGP API
changed in a few ways.  Notably, the `outputFile` property of variant
outputs is now properly implemented, and the PackageLibraryTask's
archive path now points to an AAR, and not a folder.

We account for both of those changes by updating the modern task such
that it accepts an input file, and not an input directory.

Fixes #217