Stop using ApkVariant#getOutputFile()
Although it itself isn't marked deprecated, The internal implementation
uses deprecated methods; if not de jure, it is de facto deprecated.

It appears that the only supported way to get the path of an output file
is to construct it yourself; we'll do that here, and add checks to the
integration test suite to catch any future deprecations.

Fixes #263