Split single counting task into treegen and output tasks (#321)
This is a big'un. The TL;DR is that in essence, this PR splits the single "dexcount task" into "generate PackageTree" and "print to stdout" tasks. The former does the heavy lifting of cracking open APKs, AARs, etc, and writes the PackageTree (along with the summary, chart, and formatted outputs) to disk. This new task is cacheable.

The "print to stdout" task simply reads this intermediate tree and writes all the summaries, diagnostics, etc that one would expect. This task is not cacheable, and has no outputs - it'll run all the time, regardless of the build cache.

This PR is not adequately tested - I'll add tests covering older AGP versions when handling #304.

Fixes #316